third_party/WebKit/LayoutTests/fast/shapes/parsing/parsing-shape-property-aliases-expected.txt

Test that CSS Shapes properties with the -webkit prefix alias the unprefixed properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getAliasCSSValue('shape-inside','shape-inside','rectangle(0px, 0px, 0px, 0px)') is "rectangle(0px, 0px, 0px, 0px)"
PASS getAliasCSSValue('shape-outside','shape-outside','rectangle(0px, 0px, 0px, 0px)') is "rectangle(0px, 0px, 0px, 0px)"
PASS getAliasCSSValue('shape-margin','shape-margin','10px') is "10px"
PASS getAliasCSSValue('shape-padding','shape-padding','10px') is "10px"
PASS getAliasCSSValue('shape-image-threshold','shape-image-threshold','0.5') is "0.5"
PASS successfullyParsed is true

TEST COMPLETE
